Добавлены контроллеры для работы с BusinessOperation и BusinessOperationType Добавлено предсталвение для отображения списка бизнес-операций Добавлено отображение списка бизнес-операций в клиентском приложении WPF

+++ StablingApi/Models/MoneyAccount.cs
@@ -4,6 +4,6 @@
     {
         public byte MoneyAccountId { get; set; }
         public string AccountName { get; set; }
-        public double Balance { get; set; }
+        public Decimal Balance { get; set; }
     }
 }
